[#11] refactor: usePostData 훅으로 분리
